fix(docs): address page action review feedback
