--- apps/docs/lib/docs-page-actions.clipboard.test.ts
@@ -1,33 +1,0 @@
-// @vitest-environment jsdom
-
-import { afterEach, describe, expect, it, vi } from "vitest";
-import { copyTextToClipboard } from "./docs-page-actions";
-
-describe("copyTextToClipboard", () => {
-  afterEach(() => {
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-    Reflect.deleteProperty(document, "execCommand");
-  });
-
-  it("uses the Clipboard API when available", async () => {
-    const writeText = vi.fn().mockResolvedValue(undefined);
-    vi.stubGlobal("navigator", { clipboard: { writeText } });
-
-    await expect(copyTextToClipboard("hello")).resolves.toBe(true);
-    expect(writeText).toHaveBeenCalledWith("hello");
-  });
-
-  it("falls back to a temporary textarea", async () => {
-    vi.stubGlobal("navigator", {});
-    const execCommand = vi.fn().mockReturnValue(true);
-    Object.defineProperty(document, "execCommand", {
-      configurable: true,
-      value: execCommand,
-    });
-
-    await expect(copyTextToClipboard("hello")).resolves.toBe(true);
-    expect(execCommand).toHaveBeenCalledWith("copy");
-    expect(document.querySelector("textarea")).toBeNull();
-  });
-});
